Add new Redirect to Site URL setting
Add a new setting for 'Redirect to Site URL' which will accept a URL to
use as the 'main' URL for the site.

Update SEOSettingsPart to add RedirectToSiteUrl property.

Update the SEOSettingsPartDriver to add validation for the new
RedirectToSiteUrl property to ensure that it is compatible with the
'Redirect' property. This entails checking that if redirect to www is set
then we also have a URL that contains 'www' and vice versa, preventing
redirect loops.

Update the editor template for the SEOSettingsPart to add the new field.

Update the ConsistenUrlFilter to add another step which checks for and
redirects to the new RedirectToSiteUrl if necessary.